5/8

+++ bbt/Logger/Logger.cc
@@ -10,10 +10,6 @@
 #include "bbt/config/GlobalConfig.hpp"
 using namespace bbt::log;
 using namespace bbt::config;
-//#define YNET_LOG_BUFFER
-
-
-
 
 Logger* Logger::GetInstance()
 {
@@ -170,7 +166,7 @@ void Logger::Log(LOGLEVEL level ,const std::string str)
 
     snprintf(log, 35, "[%4d%02d%02d %02d:%02d:%02d.%06ld]",
                     tm_time->tm_year + 1900, tm_time->tm_mon + 1, tm_time->tm_mday,
-                    tm_time->tm_hour, tm_time->tm_min, tm_time->tm_sec, (int)dis_millseconds);
+                    tm_time->tm_hour, tm_time->tm_min, tm_time->tm_sec, dis_millseconds);
     
 
     switch (level)
@@ -255,11 +251,17 @@ std::string Logger::GetLogName()
 
     snprintf(name, namelen, "%4d%02d%02d_%02d_%02d_%02d_%06ld.log",
                     tm_time->tm_year + 1900, tm_time->tm_mon + 1, tm_time->tm_mday,
-                    tm_time->tm_hour, tm_time->tm_min, tm_time->tm_sec, (int)dis_millseconds);
+                    tm_time->tm_hour, tm_time->tm_min, tm_time->tm_sec, dis_millseconds);
     return std::string(name,strlen(name));
 }
 
-
+/**
+ * @brief 格式化输出 + 带颜色(用于命令行输出调试)
+ * 
+ * @param fmt 
+ * @param ... 
+ * @return std::string 
+ */
 std::string bbt::log::format(const char* fmt, ...)
 {
     char        data[ARRAY_SIZE];
